Don't risk monkeypatching rspec-rails, not needed
